block: Switch to 64-bit bl.max_transfer
This change has no semantic impact: all drivers either leave the
value at 0 (no inherent 32-bit limit is still translated into
fragmentation below 2G; see the previous patch for that audit), or
set it to a value less than 2G.  However, switching to a larger
type and enforcing the 2G cap at the block layer makes it easier
to audit specific drivers for their robustness to larger sizing,
by letting them specify a value larger than INT_MAX if they have
been audited to be 64-bit clean.

Message-Id: <20181115020334.1189829-6-eblake@redhat.com>
Based-on: <20181114210548.1098207-1-eblake@redhat.com>
Based-on: <20181101182738.70462-1-vsementsov@virtuozzo.com>

+++ include/block/block_int.h
@@ -549,9 +549,11 @@ typedef struct BlockLimits {
 
     /* Maximal transfer length in bytes.  Need not be power of 2, but
      * must be multiple of opt_transfer and bl.request_alignment, or 0
-     * for no 32-bit limit.  For now, anything larger than INT_MAX is
-     * clamped down. */
-    uint32_t max_transfer;
+     * for no 32-bit limit.  The block layer fragments all actions
+     * below 2G, so setting this value to anything larger than INT_MAX
+     * implies that the driver has been audited for 64-bit
+     * cleanness. */
+    uint64_t max_transfer;
 
     /* memory alignment, in bytes so that no bounce buffer is needed */
     size_t min_mem_alignment;
